🐛 fix(predictor): 修复参数类型和清理调试输出
1. 修复 get_param<int> → get_param<int64_t> 类型错误
   - armor_motion.cpp: max_reject 参数
   - spin_motion.cpp: max_reject 参数 (两处)

2. 注释掉调试输出
   - vehicle_model.cpp: 多装甲板调试信息
   - armor_observer.cpp: PnP 和 DoubleZ 调试信息
   - predictor_node.cpp: 输入装甲板数量

+++ aimer/auto_aim/predictor/enemy_model/motion/armor_motion.cpp
@@ -68,7 +68,7 @@ void FilterThread::update(const ArmorData& armor, double timestamp) {
 
     // 读取门限参数
     double chi2_threshold = runtime_param::get_param<double>("AutoAim.Predictor.EKF.Gating.chi2_threshold");
-    int max_reject = runtime_param::get_param<int>("AutoAim.Predictor.EKF.Gating.max_reject");
+    int64_t max_reject = runtime_param::get_param<int64_t>("AutoAim.Predictor.EKF.Gating.max_reject");
     double q_scale_increase = runtime_param::get_param<double>("AutoAim.Predictor.EKF.Gating.q_scale_increase");
     double q_scale_decay = runtime_param::get_param<double>("AutoAim.Predictor.EKF.Gating.q_scale_decay");
